+++ modules.tf
@@ -42,9 +42,15 @@ module "function_app" {
   app_service_plan_id               = module.app_service_plan.app_service_plan_id
   function_language_for_linux       = var.function_language_for_linux
   function_app_application_settings = var.function_app_application_settings
+  function_app_version              = var.function_app_version
 
   application_insights_instrumentation_key = var.application_insights_instrumentation_key
   application_insights_type                = var.application_insights_type
+
+  identity_type = var.identity_type
+  identity_ids  = var.identity_ids
+
+  os_type = var.function_app_os_type
 
   extra_tags = merge(var.extra_tags, local.default_tags)
   application_insights_extra_tags = merge(