+++ versions.tf
@@ -2,6 +2,6 @@ terraform {
   required_version = ">= 0.12"
 
   required_providers {
-    azurerm = ">= 2.0"
+    azurerm = ">= 2.21.0"
   }
 }